[Refactor] #23 도메인 모델 패키지 분리
