+++ DaOnGil/data/src/main/java/kr/tekit/lion/data/repository/SigunguCodeRepositoryImpl.kt
@@ -3,8 +3,8 @@ package kr.tekit.lion.data.repository
 import kr.tekit.lion.data.database.entity.toDomainModel
 import kr.tekit.lion.data.database.entity.toEntity
 import kr.tekit.lion.data.datasource.SigunguCodeDatasource
-import kr.tekit.lion.domain.model.SigunguCode
-import kr.tekit.lion.domain.model.SigunguCodeList
+import kr.tekit.lion.domain.model.area.SigunguCode
+import kr.tekit.lion.domain.model.area.SigunguCodeList
 import kr.tekit.lion.domain.repository.SigunguCodeRepository
 import javax.inject.Inject
