+++ DaOnGil/domain/src/main/java/kr/tekit/lion/domain/model/area/SigunguCode.kt
@@ -1,4 +1,4 @@
-package kr.tekit.lion.domain.model
+package kr.tekit.lion.domain.model.area
 
 data class SigunguCode (
     val areaCode: String,
