+++ DaOnGil/domain/src/main/java/kr/tekit/lion/domain/repository/AreaCodeRepository.kt
@@ -1,7 +1,7 @@
 package kr.tekit.lion.domain.repository
 
-import kr.tekit.lion.domain.model.AreaCode
-import kr.tekit.lion.domain.model.AreaCodeList
+import kr.tekit.lion.domain.model.area.AreaCode
+import kr.tekit.lion.domain.model.area.AreaCodeList
 
 interface AreaCodeRepository {
     suspend fun getAreaCodeByName(areaName: String): String?
