+++ DaOnGil/domain/src/main/java/kr/tekit/lion/domain/repository/KorWithRepository.kt
@@ -1,6 +1,6 @@
 package kr.tekit.lion.domain.repository
 
-import kr.tekit.lion.domain.model.AreaCode
+import kr.tekit.lion.domain.model.area.AreaCode
 
 interface KorWithRepository {
     suspend fun getAreaCodeInfo(): List<AreaCode>
