+++ DaOnGil/domain/src/main/java/kr/tekit/lion/domain/repository/PlaceRepository.kt
@@ -1,10 +1,10 @@
 package kr.tekit.lion.domain.repository
 
 import kotlinx.coroutines.flow.Flow
-import kr.tekit.lion.domain.model.ListSearchOption
-import kr.tekit.lion.domain.model.ListSearchResultList
-import kr.tekit.lion.domain.model.MapSearchOption
-import kr.tekit.lion.domain.model.MapSearchResultList
+import kr.tekit.lion.domain.model.search.ListSearchOption
+import kr.tekit.lion.domain.model.search.ListSearchResultList
+import kr.tekit.lion.domain.model.search.MapSearchOption
+import kr.tekit.lion.domain.model.search.MapSearchResultList
 import kr.tekit.lion.domain.model.Result
 
 interface PlaceRepository {
