+++ DaOnGil/domain/src/main/java/kr/tekit/lion/domain/repository/SigunguCodeRepository.kt
@@ -1,7 +1,7 @@
 package kr.tekit.lion.domain.repository
 
-import kr.tekit.lion.domain.model.SigunguCode
-import kr.tekit.lion.domain.model.SigunguCodeList
+import kr.tekit.lion.domain.model.area.SigunguCode
+import kr.tekit.lion.domain.model.area.SigunguCodeList
 
 interface SigunguCodeRepository {
     suspend fun getSigunguCodeByVillageName(villageName: String): String?
