+++ DaOnGil/presentation/src/main/java/kr/tekit/lion/presentation/main/model/ListSearchUIModel.kt
@@ -1,7 +1,6 @@
 package kr.tekit.lion.presentation.main.model
 
-import kr.tekit.lion.domain.model.ListSearchResultList
-import kr.tekit.lion.domain.model.MapSearchResultList
+import kr.tekit.lion.domain.model.search.ListSearchResultList
 import kr.tekit.lion.presentation.R
 
 sealed class ListSearchUIModel(val id: Int)
